requirements.txt
asgiref==3.8.1
autopep8==1.6.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
dj-database-url==0.5.0
Django==5.1
django-environ==0.8.1
django-filter==24.2
django-jalali==6.2.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==20.1.0
idna==3.7
jalali_core==1.0.0
jdatetime==5.0.0
kavenegar==1.1.2
Markdown==3.6
pillow==10.3.0
pycodestyle==2.8.0
PyJWT==2.8.0
pypng==0.20220715.0
PyQRCode==1.2.1
pytz==2021.3
qrcode==7.4.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
whitenoise==5.3.0
